# fix(cache): improve perf on values query in sqlite cache