I've used the node-os-utils to keep track of the CPU usage. However, during the sqlite reading duration, the cpu usage did not log anything because the main thread was busy with sqlite. 

The test cache database has a size of 6Go (approximately 60 000 000 values).
Here we have the output without the fix:
```
2022-10-26T15:10:32.161Z || CPU:  11.06
2022-10-26T15:10:32.664Z || CPU:  7.96
2022-10-26T15:10:33.163Z || CPU:  12.89
2022-10-26T15:10:33.665Z || CPU:  14.16
2022-10-26T15:10:34.165Z || CPU:  8.14
2022-10-26T15:10:34.667Z || CPU:  7.59
2022-10-26T15:10:35.166Z || CPU:  6.36
2022-10-26T15:10:37.222Z || CPU:  16.58
2022-10-26T15:10:37.223Z || CPU:  19.16
retrieve: 3.747s
North Console sent 10000 values.
remove: 12.186ms
2022-10-26T15:10:40.983Z || CPU:  19.42
```

And with the fix:
```
2022-10-26T15:09:38.694Z || CPU:  8.12
2022-10-26T15:09:39.197Z || CPU:  6.37
2022-10-26T15:09:39.699Z || CPU:  7.64
2022-10-26T15:09:40.201Z || CPU:  7.64
2022-10-26T15:09:42.217Z || CPU:  15.71
2022-10-26T15:09:42.218Z || CPU:  18.02
retrieve: 18.738ms
North Console sent 10000 values.
remove: 11.992ms
2022-10-26T15:09:43.219Z || CPU:  12.06
2022-10-26T15:09:43.720Z || CPU:  10.32
2022-10-26T15:09:44.221Z || CPU:  12.83
```

I think the cpu usage take the average cpu which is overall lower with the fix since the cpu is used less in the fix case.

The retrieve time is greatly improved (from 3,7s to 19ms)

About the fix:
I've removed the `ORDER BY timestamp `clause which is not useful since we already have a FIFO logic and the cache is now separated for each North (thank to the refactor !)

I've also remove the iterate method to use the all method, [slightly better according to the documentation](https://github.com/WiseLibs/better-sqlite3/blob/master/docs/api.md#iteratebindparameters---iterator).